% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{add_annotation}
\alias{add_annotation}
\title{Annotate tables}
\usage{
add_annotation(path_to_annotation_table, table, keep)
}
\arguments{
\item{path_to_annotation_table}{path to a .gtf table containing the annotation.}

\item{table}{a table to annotate. Must contain the col. chr/start /stop}

\item{keep}{annotation to take into consideration.}
}
\value{
the table to annotad with an extra list col: annotation
}
\description{
Add a column to the table with the gene name.
}
