% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_meth.R
\name{add_count_colnames_to_covariates_tables}
\alias{add_count_colnames_to_covariates_tables}
\title{Extract the count of a count table}
\usage{
add_count_colnames_to_covariates_tables(covariate_table)
}
\arguments{
\item{covariate_table}{A data frame with the ID col.}
}
\value{
A data.frame with the same column as the input plus a col with the
corresponding names in the count_table.
}
\description{
Extract the count of a count table
}
