% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DMRs.R
\name{add_DMRs_to_set}
\alias{add_DMRs_to_set}
\title{Add the DMRs as a list_col to the set table.
nb. the coef are hard coded for the current usage but can easily be changed.}
\usage{
add_DMRs_to_set(set, full_tables, formula, p_cut, index = NULL, method = "GLM")
}
\arguments{
\item{set}{the table created with prepare_balanced_k_set().}

\item{full_tables}{the covariate and count tables created with prepare_ws_datasets()}

\item{formula}{the formula for the model}

\item{p_cut}{the p_value threshold on which to keep the DMR. (mostly used to keep object size decent)}

\item{index}{a vector of intex to retain the value of the fit. THought for the CV to keep the value of the DMR in the full set}

\item{method}{GLM or QL}
}
\value{
A nested DF with DMRs for each subset
}
\description{
Add the DMRs as a list_col to the set table.
nb. the coef are hard coded for the current usage but can easily be changed.
}
\examples{
data(full_300_ws_tables_1000)
set <- prepare_balanced_k_set(6, full_300_ws_tables_1000, "group")
set2 <- add_DMRs_to_set(set, full_300_ws_tables_1000, "~group + age_group", p_cut = 1)
}
