% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitochandria.R
\name{compute_mean_cpm_mito}
\alias{compute_mean_cpm_mito}
\title{Compute mean cpm for mito and non-mito genes}
\usage{
compute_mean_cpm_mito(full_300_ws_tables, path_to_annotation_table)
}
\arguments{
\item{full_300_ws_tables}{full tables with count and cov tables}

\item{path_to_annotation_table}{path to annotated table stored in exdata}
}
\value{
A tibble with covariate and mean cpm for mito genes and non-mito genes
}
\description{
Compute mean cpm for mito and non-mito genes
}
\examples{
\dontrun{
path_to_annotation_table <- "inst/extdata/crocuta_liftoff_Hhy_ASM300989v1_addPromoter.gtf"
load("data-raw/data/full_300_ws_tables.rda")

out <- compute_mean_cpm_mito(full_300_ws_tables, path_to_annotation_table) ## TO run on the server as the files is around 30 GB
}
}
