% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitochandria.R
\name{compute_mean_cpm_mito_rna}
\alias{compute_mean_cpm_mito_rna}
\title{Compute mean cpm for mito and non-mito genes RNA}
\usage{
compute_mean_cpm_mito_rna(RNA_table)
}
\arguments{
\item{RNA_table}{full tables with RNA count and cov tables}
}
\value{
A tibble with covariate and mean cpm for mito genes and non-mito genes expression
}
\description{
Compute mean cpm for mito and non-mito genes RNA
}
\examples{
\dontrun{
load("data-raw/data/RNA_table.rda")
mito_table_RNA <-compute_mean_cpm_mito_rna(RNA_table)
}
}
