% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DMR_and_annotated_DMR_tables.R
\name{create_annot_DMRs}
\alias{create_annot_DMRs}
\title{Make a long version of the annotated DMR_group_table}
\usage{
create_annot_DMRs(DMR_table)
}
\arguments{
\item{DMR_table}{The DMR table}
}
\value{
A long format tibble with annotated DMR
}
\description{
remove the intergenic region and keep 1 row per annotation so 1 DMR can represent
many rows.
}
