% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DMR_and_annotated_DMR_tables.R
\name{create_DMR_group_table}
\alias{create_DMR_group_table}
\title{Create a clean table with DMRs from the full set with information about CV subsets.
return the accuracy, the f1-score, the younden J-statistic, the recall and precision rates.}
\usage{
create_DMR_group_table(
  model_1_one_fold,
  model_1_six_fold,
  path_to_annotation_table,
  loc_table
)
}
\arguments{
\item{model_1_one_fold}{output of the fitting of model 1 fold}

\item{model_1_six_fold}{output of the fitting of model 6 fold}

\item{path_to_annotation_table}{path to annotation table stored in int/exdata}

\item{loc_table}{internal DMR table with indexes to merge}
}
\value{
A tibble with DMR with full DMRs information
}
\description{
Create a clean table with DMRs from the full set with information about CV subsets.
return the accuracy, the f1-score, the younden J-statistic, the recall and precision rates.
}
\examples{
\dontrun{
data(model_1_one_fold)
data(model_1_six_fold)
path_to_annotation_table <- "~/Documents/Weyrich23/inst/extdata/crocuta_liftoff_Hhy_ASM300989v1_addPromoter.gtf"
data(loc_table)
create_DMR_group_table(model_1_one_fold, model_1_six_fold, path_to_annotation_table, loc_table)
}
}
