% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_descriptive.R
\name{describes_samples}
\alias{describes_samples}
\title{describe the samples}
\usage{
describes_samples(tables)
}
\arguments{
\item{tables}{table with read_counts and covariates}
}
\value{
a list of tibble with samples descriptiomns
}
\description{
Give the mean, n, and sd for group and age_group
}
