% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DMR_adult_young_comp}
\alias{DMR_adult_young_comp}
\title{DMR_adult_young_comp}
\format{
\code{DMR_adult_young_comp}
DMRs found in the full dataset with information about variance in the subsets in adult and young sepparetely
}
\usage{
data(DMR_adult_young_comp)
}
\description{
table with the effect size in the main DMR for young and adult fitted separately
}
\keyword{datasets}
