% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DMR_bootstrap_adult_young}
\alias{DMR_bootstrap_adult_young}
\title{DMR_bootstrap_adult_young}
\format{
\code{DMR_bootstrap_adult_young}
coeficient of corellation in 100 bootsraped windows between adult and young sepparetely
}
\usage{
data(DMR_bootstrap_adult_young)
}
\description{
table with the effect size in 1000 random  windows for young and adult fitted separately
}
\keyword{datasets}
