% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DMR_group}
\alias{DMR_group}
\title{DMR_group}
\format{
\code{DMR_group}
A nested tibble with 1 row per subset.
}
\usage{
data(DMR_group)
}
\description{
DMRs found in the full dataset with information about variance in the subsets
}
\keyword{datasets}
