% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DMR_and_annotated_DMR_tables.R
\name{.add_merged_DMR}
\alias{.add_merged_DMR}
\title{Provide summary statistic on the performance of the classifier.
return the accuracy, the f1-score, the younden J-statistic, the recall and precision rates.}
\usage{
.add_merged_DMR(DMR_full)
}
\arguments{
\item{DMR_full}{internal DMR table with indexes to merge}
}
\value{
A tibble 2 added col. block and block_num
}
\description{
Provide summary statistic on the performance of the classifier.
return the accuracy, the f1-score, the younden J-statistic, the recall and precision rates.
}
