% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_meth.R
\name{.compute_mean_read_count}
\alias{.compute_mean_read_count}
\title{create a data.frame with the mean read count / mean TMM read count per pos per group.}
\usage{
.compute_mean_read_count(design, coef, count_table)
}
\arguments{
\item{design}{A data frame with the design matrix of the fitted model}

\item{coef}{A character string specifying the coeficient of interest - need to be a col of the design df.}

\item{count_table}{A table of count, with colnames as "IDs_count".}
}
\value{
A data.frame with the mean read count per group (coef == 1) for count and count_TMM.
}
\description{
create a data.frame with the mean read count / mean TMM read count per pos per group.
}
