% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{.creat_indexes_list}
\alias{.creat_indexes_list}
\title{create a list with indexes for DMRs -robust, non-robust, full, and randoms...}
\usage{
.creat_indexes_list(DMR_group, tables)
}
\arguments{
\item{DMR_group}{tables with DMR and robustness info.}

\item{tables}{list with read count table and cov table}
}
\value{
A list with the indexes of coherent and non coherent DMR + libsize.
}
\description{
create a list with indexes for DMRs -robust, non-robust, full, and randoms...
}
