% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{.extract_standardized_lib_size}
\alias{.extract_standardized_lib_size}
\title{create a list with indexes for DMRs -robust, non-robust, full, and randoms...}
\usage{
.extract_standardized_lib_size(tables)
}
\arguments{
\item{tables}{the full_300_ws_tables}
}
\value{
A list with the indexes of coherent and non coherent DMR + libsize.
}
\description{
create a list with indexes for DMRs -robust, non-robust, full, and randoms...
}
