% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DMR_and_annotated_DMR_tables.R
\name{.identify_incoherent_DMR}
\alias{.identify_incoherent_DMR}
\title{Provide summary statistic on the performance of the classifier.
return the accuracy, the f1-score, the younden J-statistic, the recall and precision rates.}
\usage{
.identify_incoherent_DMR(DMR_CV_table)
}
\arguments{
\item{DMR_CV_table}{merged table with DMR from subsets and full dataset.}
}
\value{
A tibble 2 added col. block and block_num
}
\description{
Provide summary statistic on the performance of the classifier.
return the accuracy, the f1-score, the younden J-statistic, the recall and precision rates.
}
