% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{.rename_loc}
\alias{.rename_loc}
\title{Add annotation to loc genes}
\usage{
.rename_loc(annotation_table, loc_table)
}
\arguments{
\item{annotation_table}{A annotation data frame with "type" columns.}

\item{loc_table}{A table with the new gene names}
}
\value{
A annotation tables with changed names for loc genes with new annotation
}
\description{
Add annotation to loc genes
}
