% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig1.R
\name{.reshape_data_fig1}
\alias{.reshape_data_fig1}
\title{reshape the data for fig1.}
\usage{
.reshape_data_fig1(DMR_group)
}
\arguments{
\item{DMR_group}{table with DMR}
}
\value{
A tibble with data for fig.1
}
\description{
reshape the data for fig1.
}
