% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2_and_sup.R
\name{.reshape_data_fig2}
\alias{.reshape_data_fig2}
\title{reshape the data for fig2.}
\usage{
.reshape_data_fig2(DMR_group)
}
\arguments{
\item{DMR_group}{table with DMR}
}
\value{
A tibble with data for fig.1
}
\description{
reshape the data for fig2.
}
