% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_s2b.R
\name{.reshape_data_fig_s2b}
\alias{.reshape_data_fig_s2b}
\title{reshape the data for fig2.}
\usage{
.reshape_data_fig_s2b(tables)
}
\arguments{
\item{tables}{ttable with read_counts and covariates}
}
\value{
A tibble with data for fig.1
}
\description{
reshape the data for fig2.
}
