% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{estimate_oob}
\alias{estimate_oob}
\title{Fit random forest model trained on DMR and mean meth.}
\usage{
estimate_oob(DMR_group, set, tables, n = 1000)
}
\arguments{
\item{DMR_group}{table with DMRs for group variable}

\item{set}{the nested table containing the fitted model}

\item{tables}{the full_300_ws_tables}

\item{n}{number of bootstrap}
}
\value{
A tibble with the mean oob error accross bootstrap
}
\description{
Bootstrap X times the fit_and_summarise_RF to have a less variable
estimate of the oob error. Especially for the randomise run.
}
