% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_meth.R
\name{extract_count_matrix}
\alias{extract_count_matrix}
\title{Extract the count of a count table}
\usage{
extract_count_matrix(count_table, which = "count")
}
\arguments{
\item{count_table}{A data frame of counts. Idealy created with the
Epiena::prepare_ROI_datasets()]or Epiena::prepare_ws_datasets() from the Epiena package.  Each count column must ends with "count".
The first three col of the table shoup be the "chr", "start" and "stop".}

\item{which}{either "count" or "count_TMM". It should match the end of the count columns.}
}
\value{
A named matrix countaing the read counts.
}
\description{
Extract the count of a count table
}
