% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{extract_indexes}
\alias{extract_indexes}
\title{Extract the indexes of the DMR positions. (needed to get the correct CPM in the count_table)}
\usage{
extract_indexes(table_with_DMR, DMR_to_use, q_cut)
}
\arguments{
\item{table_with_DMR}{the DMR_train col of the set.}

\item{DMR_to_use}{The coef to use (either groupLR, or age_groupyoung)}

\item{q_cut}{The cut-off value on the FDR value.}
}
\value{
A vector of indexes
}
\description{
Extract the indexes of the DMR positions. (needed to get the correct CPM in the count_table)
}
