% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig1.R
\name{fig1}
\alias{fig1}
\title{Create fig1}
\usage{
fig1(DMR_group, save = FALSE, ...)
}
\arguments{
\item{DMR_group}{table with DMR}

\item{save}{otption to save the file, nee}

\item{...}{argument to pass to ggsave()}
}
\value{
A plot
}
\description{
Create fig1
}
\examples{
\dontrun{
data("DMR_group")
 fig1(DMR_group, save = TRUE, filename = "figures/fig1_final.png", width = 12, height = 8, units = "in", dpi = 300, bg = "white")
 fig1(DMR_group, save = TRUE, filename = "figures/fig1_final.pdf", width = 3.5, height = 2.2, units = "in", dpi = 300, bg = "white")
}


}
