% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2_and_sup.R
\name{fig2}
\alias{fig2}
\title{Create fig2}
\usage{
fig2(DMR_group, save = FALSE, sup = FALSE, ...)
}
\arguments{
\item{DMR_group}{table with DMR}

\item{save}{otption to save the file, nee}

\item{sup}{logical to create the suplementary table}

\item{...}{argument to pass to ggsave()}
}
\value{
A plot
}
\description{
Create fig2
}
\examples{
data(DMR_group)
fig2(DMR_group, sup = TRUE)
}
