% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig3.R
\name{fig3}
\alias{fig3}
\title{Create fig3}
\usage{
fig3(mito_table_DNA, mito_table_RNA, save = FALSE, ...)
}
\arguments{
\item{mito_table_DNA}{table with mean cpm for mito and non mito genes DNA}

\item{mito_table_RNA}{table with mean cpm for mito and non mito genes RNA}

\item{save}{otption to save the file, nee}

\item{...}{argument to pass to ggsave()}
}
\value{
A plot
}
\description{
Create fig3
}
