% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_s2b.R
\name{fig_s2b}
\alias{fig_s2b}
\title{Create fig1}
\usage{
fig_s2b(tables, save = FALSE, ...)
}
\arguments{
\item{tables}{table with read_counts and covariates}

\item{save}{otption to save the file, nee}

\item{...}{argument to pass to ggsave()}
}
\value{
A plot
}
\description{
Create fig1
}
\examples{
\dontrun{
data(full_300_ws_tables_1000)
fig_s2b(full_300_ws_tables_1000,save= FALSE)
}
}
