% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_s4.R
\name{fig_s4}
\alias{fig_s4}
\title{Create fig_S4}
\usage{
fig_s4(
  DMR_group,
  DMR_adult_young_comp,
  DMR_bootstrap_adult_young,
  save = FALSE,
  ...
)
}
\arguments{
\item{DMR_group}{table with DMR on the full set}

\item{DMR_adult_young_comp}{table with Adult/youg fitted separately on DMR}

\item{DMR_bootstrap_adult_young}{table with adult/young fitted separately on random windows}

\item{save}{otption to save the file, nee}

\item{...}{argument to pass to ggsave()}
}
\value{
A plot
}
\description{
Create fig_S4
}
\examples{
\dontrun{
data("DMR_group")
data("DMR_adult_young_comp")
data("DMR_bootstrap_adult_young")
fig_s4(DMR_group, DMR_adult_young_comp, DMR_bootstrap_adult_young, save = TRUE, filename = "figures/figs4_final2.png", my_plot, device = "png", width = 12, units = "in", height = 8, dpi = 300, bg = "white")
}
}
