% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{find_annotation_type}
\alias{find_annotation_type}
\title{find annotation type}
\usage{
find_annotation_type(annotation_table = NULL, path_to_annotation_table = NULL)
}
\arguments{
\item{annotation_table}{A annotation data frame with "type" columns.}

\item{path_to_annotation_table}{the annotation table}
}
\value{
the vector with the diffent unique annotation types
}
\description{
find annotation type
}
