% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_meth.R
\name{find_coef}
\alias{find_coef}
\title{Help function to find the possible coef given a certain formula.}
\usage{
find_coef(covariate_table, formula)
}
\arguments{
\item{covariate_table}{A data frame with the ID col. and other covariate.}

\item{formula}{formula on which to fit the data. The variable used in the
formula should be present in the covariate_table.}
}
\value{
A vector with coef names.
}
\description{
Help function to find the possible coef given a certain formula.
}
