% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{fit_and_summarise_RF}
\alias{fit_and_summarise_RF}
\title{Fit random forest model trained on DMR and mean meth.}
\usage{
fit_and_summarise_RF(DMR_group, set, tables)
}
\arguments{
\item{DMR_group}{table with DMRs for group variable}

\item{set}{the nested table containing the fitted model}

\item{tables}{the full_300_ws_tables}
}
\value{
A tibble with oob error and other variables
}
\description{
run on the list of indexes and return a tibble
}
\examples{
\dontrun{
load("data-raw/data/full_300_ws_tables.rda")
data(DMR_group)
data(model_1_one_fold)
fit_and_summarise_RF(DMR_group, model_1_one_fold, full_300_ws_tables)
}
}
