% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_meth.R
\name{fit_edgeR}
\alias{fit_edgeR}
\title{Fit a negative Binomial model per position.}
\usage{
fit_edgeR(
  count_table,
  covariate_table,
  formula = "~ group + age_group",
  coef = "groupLR",
  mean_read = FALSE,
  method = "GLM"
)
}
\arguments{
\item{count_table}{A table of count, with colnames as "IDs_count".}

\item{covariate_table}{A data frame with the ID col. and other covariate.}

\item{formula}{A formula on which to fit the data. The variable used in the
formula should be present in the covariate_table.}

\item{coef}{A coeficient of interest on which the statistics will be computed.}

\item{mean_read}{logical to specify if the mean read count per group should be computed. FALSE for num or fit with multiple covariate.}

\item{method}{Specification of the method to use to fit the data quasi likelihood = "QL or GLM.  Default GLM}
}
\value{
A list with the fitted objects and a data.frame with the results.
}
\description{
Fit a negative Binomial model per position.
}
\examples{
data(full_300_ws_tables_1000)
count_table <- full_300_ws_tables_1000$count_table
covariate_table <- full_300_ws_tables_1000$covariate_table
fit_edgeR(count_table, covariate_table, formula = " ~ group + age_group",
          coef = "groupLR", mean_read = FALSE, method = "GLM")
}
