% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_meth.R
\name{fit_model}
\alias{fit_model}
\title{Fit a negative Binomial model per position.}
\usage{
fit_model(count_table, covariate_table, formula, method = "GLM")
}
\arguments{
\item{count_table}{A table of count, with colnames as "IDs_count".}

\item{covariate_table}{A data frame with the ID col. and other covariate.}

\item{formula}{A formula on which to fit the data. The variable used in the}

\item{method}{Specification of the method to use to fit the data quasi likelihood = "QL or GLM}
}
\value{
A fitted edgeR model.
}
\description{
Fit a negative Binomial model per position.
}
