% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_300_ws_tables_1000}
\alias{full_300_ws_tables_1000}
\title{full_300_ws_tables_1000}
\format{
\code{full_300_ws_tables_1000}
A list of 2 data-frames
\describe{
\item{count_table}{data-frame with read count per samples per window}
\item{covariate_table}{data-frame with information about the samples}
}
}
\usage{
data(full_300_ws_tables_1000)
}
\description{
A subset of the base table to run the analysis. It is composed of 2 data-frame.
The first consist of the read count, the normalised read count (using TMM),
for each samples and each windows of 300bp. This subset consist of the first
1000 windows. In addition we have the \link{chr}, \link{start}, amd \link{stop} columns
refering to the scaffold, the first and last position of the windows.
}
\details{
A subset of data
}
\keyword{datasets}
