% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DMR_and_annotated_DMR_tables.R
\name{get_annotation_type}
\alias{get_annotation_type}
\title{Get the annotation type}
\usage{
get_annotation_type(annotated_DMR)
}
\arguments{
\item{annotated_DMR}{The annotated DMR table}
}
\value{
A tibble with gene annotation including intron
}
\description{
Add intron information. Region are considered intron if they overlap with gene and no exon.
}
