% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DMRs.R
\name{get_DMRs}
\alias{get_DMRs}
\title{Fit a negbin model using EdgeR to detect DMR.}
\usage{
get_DMRs(tables, formula, coef, p_cut, index = NULL, method = "GLM")
}
\arguments{
\item{tables}{the covariate and count tables created with prepare_ws_datasets()}

\item{formula}{the formula for the model}

\item{coef}{the coeficient to extract the DMR for (can be multiple).}

\item{p_cut}{the p_value threshold on which to keep the DMR. (mostly used to keep object size decent)}

\item{index}{a vector of intex to retain the value of the fit. THought for the CV to keep the value of the DMR in the full set}

\item{method}{GLM or QL}
}
\value{
A tibble with DMRs
}
\description{
nb. It uses fit_edgeR that is part of the epiena package developped to work with this datasets.
and is a wrapper around the edgeR functions.
}
