% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{load_and_shape_annotation_table}
\alias{load_and_shape_annotation_table}
\title{load and shape the annotation table}
\usage{
load_and_shape_annotation_table(
  path_to_annotation_table,
  keep = "all",
  output = "data.frame"
)
}
\arguments{
\item{path_to_annotation_table}{path to a .gtf table containing the annotation.}

\item{keep}{the typoe of annotation to take into account.}

\item{output}{The format of the output: either "data.frame" or "GRanges"}
}
\value{
A GRange of a data.frame  with the genomic coordinates, the geneID and the type.
}
\description{
load and shape the annotation table
}
\examples{
\dontrun{
x <- load_and_shape_annotation_table(path_to_annotation_table, keep = "all", output = "data.frame")
}

}
