% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loc_table}
\alias{loc_table}
\title{loc_table}
\format{
\code{loc_table}
A tibble with loc names and gene names
}
\usage{
data(loc_table)
}
\description{
table with loc names and corresponding annotation.
}
\keyword{datasets}
