% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_table_DNA}
\alias{mito_table_DNA}
\title{mito_table_DNA}
\format{
\code{mito_table_DNA}
A tibble with mean cpm for mito and non mito genes per ID
}
\usage{
data(mito_table_DNA)
}
\description{
table with the mean cpm for mito and non-mito genes
fitted on a server with the same code as in the vignette (big files)
}
\keyword{datasets}
