% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_table_RNA}
\alias{mito_table_RNA}
\title{mito_table_RNA}
\format{
\code{mito_table_RNA}
A tibble with mean cpm for mito and non mito genes per ID
}
\usage{
data(mito_table_RNA)
}
\description{
table with the mean cpm for mito and non-mito genes RNA
}
\keyword{datasets}
