% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_1_one_fold}
\alias{model_1_one_fold}
\title{model_1_one_fold}
\format{
\code{model_1_one_fold}
A nested tibble with 1 row and 4 variables
\describe{
\item{set_ID}{the identifier of the set}
\item{data_test}{A list column of  data-frame with the covariate table of the samples composing the set (all samples), The random forrest is trained on this set}
\item{data_train}{The same list column as data_test}
\item{DMR_train}{A list column of data-frame with the result of the neg-bin fit for the windows with significant difference between HR and LR}
}
}
\usage{
data(model_1_one_fold)
}
\description{
The full set with its DMRs. It is organised as a nested tibble with 4 variables.
Used for the fitting of the models and the Random Forest.
Contains the result of the negative binomial model for the full dataset. It contains
one row only containing all the samples in both the train set and test set
variable. It have the same structure than the \link{model_1_six_fold} table.
}
\keyword{datasets}
