% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_1_six_fold}
\alias{model_1_six_fold}
\title{model_1_six_fold}
\format{
\code{model_1_six_fold}
A nested tibble with 1 row per subsets (6) and 4 variables
\describe{
\item{set_ID}{the identifier of the set}
\item{data_test}{A list column of  data-frame with the covariate table of the samples composing the subset}
\item{data_train}{A list column of data-frame with the covariate table of the samples removed from the subset}
\item{DMR_train}{A list column of data-frame with the result of the neg-bin fit for the windows with significant difference between HR and LR}
}
}
\usage{
data(model_1_six_fold)
}
\description{
The six subsets with theirs DMRs. It is organised as a nested tibble with 4 variables,
and one row per subsets. Used for the robustness analysis of the DMRs.
Contains the result of the negative binomial model for the six subsets.
It contains six row each containing the samples used for the fitting in the
train set and the one removed from the analysis in the test set. The DMRs
are columns contains the position passing the cutof and the identified in
the full set.
It have the same structure than the \link{model_1_one_fold} table.
}
\keyword{datasets}
