% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{pivote_and_trim_table}
\alias{pivote_and_trim_table}
\title{Merge the information about the classes (age / rank) and the TMM count numbers}
\usage{
pivote_and_trim_table(
  cov_table,
  count_table,
  index,
  covariate_from_formula = c("group", "age_group")
)
}
\arguments{
\item{cov_table}{covariate tables containing group and age_group.}

\item{count_table}{the count_tables for the specific ID and indexes}

\item{index}{the indexes of the DMR position to keep track in later steps.}

\item{covariate_from_formula}{the covariate used to fit the model to pass as covariate for the RF}
}
\value{
A tibble with CPM and covariate from DMRs postion
}
\description{
for the specific positions.
}
