% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DMRs.R
\name{prepare_balanced_k_set}
\alias{prepare_balanced_k_set}
\title{Create a data frame with list-col with data_train and data_test.}
\usage{
prepare_balanced_k_set(k, tables, variable_to_balance)
}
\arguments{
\item{k}{the number of folds}

\item{tables}{the covariate and count tables created with prepare_ws_datasets()}

\item{variable_to_balance}{the variable name that need to be present in both sets.}
}
\value{
A nested tibble with K row
}
\description{
Keeping the variable of interest (group) balanced.
}
\examples{
data(full_300_ws_tables_1000)
prepare_balanced_k_set(6, full_300_ws_tables_1000, "group")

}
