% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DMRs.R
\name{reduce_count_table}
\alias{reduce_count_table}
\title{Filter the count_table to keep ONLY the count and count_TMM for the IDs in the train set.}
\usage{
reduce_count_table(train_set, tables)
}
\arguments{
\item{train_set}{the talbe with the IDs in the train set.}

\item{tables}{the covariate and count tables created with prepare_ws_datasets()}
}
\value{
A nested tibble with K row
}
\description{
Filter the count_table to keep ONLY the count and count_TMM for the IDs in the train set.
}
