% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adult_young_comp_table.R
\name{refit_DMR_adult_young_separately}
\alias{refit_DMR_adult_young_separately}
\title{Create Young / adult comparison table}
\usage{
refit_DMR_adult_young_separately(
  model_1,
  full_tables,
  path_to_annotation_table
)
}
\arguments{
\item{model_1}{fitted model 1 on the full datasets}

\item{full_tables}{full 300 ws tables}

\item{path_to_annotation_table}{annotation table}
}
\value{
A data frame with effect size for DMR position in both adult and young
}
\description{
Create Young / adult comparison table
}
\examples{
\dontrun{
data("model_1_one_fold")
load("../../Downloads/Weyrich_EpiRank/full_300_ws_tables.rda")
path_to_annotation_table <- "../../Downloads/Weyrich_EpiRank/crocuta_liftoff_Hhy_ASM300989v1_addPromoter.gtf"
DMR_adult_young_comp  <- refit_DMR_adult_young_separately(model_1_one_fold, full_300_ws_tables, path_to_annotation_table)
save(DMR_adult_young_comp, file = "data/DMR_adult_young_comp.rda")
}

}
