% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models_meth.R
\name{reshape_edge_fit}
\alias{reshape_edge_fit}
\title{Reshape the outcome of the edgeR fit.}
\usage{
reshape_edge_fit(fit, coef, method = "GLM")
}
\arguments{
\item{fit}{A edgeR-fit object.}

\item{coef}{A coeficient of interest on which the statistics will be computed.}

\item{method}{GLM or PQL}
}
\value{
A table with the pvalue, fdr, and fitting results for the given coef.
}
\description{
code need to be
}
