% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{reshape_tables_for_RF}
\alias{reshape_tables_for_RF}
\title{Create the input table for the RF models. Combine information about the
samples and the CPM at the DMR positions.}
\usage{
reshape_tables_for_RF(
  table_to_turn,
  count_table_full,
  index,
  covariate_from_formula = c("group", "age_group")
)
}
\arguments{
\item{table_to_turn}{either the train_data or the test_data}

\item{count_table_full}{the full count table in which the CPM can be extracted.}

\item{index}{the indexes of the DMR position to keep track in later steps.}

\item{covariate_from_formula}{the covariate used to fit the model to pass as covariate for the RF}
}
\value{
A tibble trimmed for the index position and for the ID to predict
}
\description{
Create the input table for the RF models. Combine information about the
samples and the CPM at the DMR positions.
}
