% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{rf_on_lib_size}
\alias{rf_on_lib_size}
\title{Fit a random forest model trained only on the ~ mean methylation.}
\usage{
rf_on_lib_size(tables)
}
\arguments{
\item{tables}{the full_300_ws_tables}
}
\value{
A tibble with oob error and other variables
}
\description{
Fit a random forest model trained only on the ~ mean methylation.
}
\examples{
\dontrun{
data(full_300_ws_tables_1000)
rf_on_lib_size(full_300_ws_tables_1000)
}
}
