% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forrest.R
\name{run_rf_on_set}
\alias{run_rf_on_set}
\title{Create the input table for the RF models. Combine information about the
samples and the CPM at the DMR positions.}
\usage{
run_rf_on_set(
  set,
  count_table_full,
  variable_to_test,
  q_cut,
  DMR_to_use,
  index = NULL,
  covariate_from_formula = c("group", "age_group")
)
}
\arguments{
\item{set}{The full set including the col: data_test, data_train; DMR_train.}

\item{count_table_full}{the full count table in which the CPM can be extracted.}

\item{variable_to_test}{The variable to be classified. Either group or age_group.}

\item{q_cut}{The cut-off value on the FDR scale.}

\item{DMR_to_use}{The coef of the DMR used to predict the variable_to_test.}

\item{index}{vector of index for targeted DMRs. Default to null, the indexes will be computed from the DMR_to_use and the q_cut}

\item{covariate_from_formula}{the covariate used to fit the model to pass as covariate for the RF}
}
\value{
A tibble with the summary results of the fitted RF.
}
\description{
Create the input table for the RF models. Combine information about the
samples and the CPM at the DMR positions.
}
