% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2_and_sup.R
\name{fig2}
\alias{fig2}
\title{Create fig2}
\usage{
fig2(DMR_group, save = FALSE, sup = FALSE, ...)
}
\arguments{
\item{DMR_group}{table with DMR}

\item{save}{otption to save the file, nee}

\item{sup}{logical to create the suplementary table}

\item{...}{argument to pass to ggsave()}
}
\value{
A plot
}
\description{
Create fig2
}
\examples{
\dontrun{
data(DMR_group)
fig2(DMR_group, save = TRUE, sup = FALSE, filename = "figures/fig2_final.png", width = 8, height = 8, units = "in", dpi = 300, bg = "white")
fig2(DMR_group, save = TRUE, sup = TRUE, filename = "figures/figS3_final.png", width = 8, height = 14, units = "in", dpi = 300, bg = "white")
}
}
